// apps/api/prisma/@generated/user/output/user-max-aggregate.output.ts
import { Field, HideField, ObjectType } from '@nestjs/graphql';

@ObjectType()
export class UserMaxAggregate {
  @Field(() => String, { nullable: true })
  id?: string;

  @Field(() => Date, { nullable: true })
  createdAt?: Date | string;

  @Field(() => Date, { nullable: true })
  updatedAt?: Date | string;

  @Field(() => String, { nullable: true })
  email?: string;

  @Field(() => String, { nullable: true })
  name?: string;

  @HideField()
  password?: string;

  @Field(() => Date, { nullable: true })
  emailConfirmedAt?: Date | string;

  @Field(() => String, { nullable: true })
  confirmationCode?: string;

  @Field(() => String, { nullable: true })
  roleId?: string;
}